temp

+++ ocl_project/BicgStabHost.cs
@@ -3,8 +3,8 @@ using System.Diagnostics;
 
 using static Solvers.Shared;
 
-using Real = float;
-using VectorReal = float[];
+using Real = double;
+using VectorReal = double[];
 
 namespace Solvers.Host
 {
